+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Improved
+- **dx**: `$.clear`, `$.fire`, `$.on`, `$.once` are now typed with `Element` rather than `HTMLElement` to ensure accuracy
+- **dx**: `$.query` and `$.queryAll` are now typed with `Node` to ensure accuracy
+
 ## [0.41.1] - 2025-06-29
 This release refines Atomic’s DOM tooling for broader compatibility and better developer experience